List Riemann as a dev dependency only
List Riemann as a dev dependency only to prevent it from being included
in the uberjar. It is not necessary because the Riemann jar will be
present from the Riemann installation.

Removing Riemann from the uberjar will help shrink the size of the
uberjar, and will allow it to be used with other versions of Riemann. If
Riemann is left in the uberjar, the classes in the plugin will override
those in the Riemann jar itself.